Expand abilities to u16
(referenced from: https://github.com/rh-hideout/pokeemerald-expansion/commit/9ee5db6a967f44c16b186242c36e29f08a542611)

+++ src/pokemon.c
@@ -3776,7 +3776,7 @@ u8 GetMonsStateToDoubles(void)
     return (aliveCount > 1) ? PLAYER_HAS_TWO_USABLE_MONS : PLAYER_HAS_ONE_USABLE_MON;
 }
 
-u8 GetAbilityBySpecies(u16 species, bool8 abilityNum)
+u16 GetAbilityBySpecies(u16 species, bool8 abilityNum)
 {
     if (abilityNum)
         gLastUsedAbility = gSpeciesInfo[species].abilities[1];
@@ -3786,7 +3786,7 @@ u8 GetAbilityBySpecies(u16 species, bool8 abilityNum)
     return gLastUsedAbility;
 }
 
-u8 GetMonAbility(struct Pokemon *mon)
+u16 GetMonAbility(struct Pokemon *mon)
 {
     u16 species = GetMonData(mon, MON_DATA_SPECIES, NULL);
     u8 abilityNum = GetMonData(mon, MON_DATA_ABILITY_NUM, NULL);
